Change SHARP dust sensor to DSM501A sensor
After received and tested the SHARP sensor, I've found the sensor is
extremely unstable. Maybe the digital signal from arduino is not right
or... for some reason the signal itself is not stable enough. So I
bought a korean-made DSM501A sensor and use the offical library from
TAOBAO retailer to refine the program.